scraper/requirements.txt
# =========================================
# EXPENSE TRACKER - SCRAPER DEPENDENCIES
# =========================================

# Web scraping
playwright>=1.40.0

# Database client
supabase>=2.3.0

# Environment variables
python-dotenv>=1.0.0

# Data processing (sin versión fija para wheels precompilados)
pandas>=2.0.0

# Date/time utilities
python-dateutil>=2.8.2

# PDF parsing
pypdf2>=3.0.0
pdfplumber>=0.10.0
